requirements.txt
openai>=1.35.0
sounddevice>=0.4.6
numpy>=1.26.0
pynput>=1.7.6
python-dotenv>=1.0.1
pyperclip>=1.8.2 